feat: update logic to save check-entity - json 파일에서 점검 정보(pm10 = 0, pm2.5 = 0)를 check-entity에 저장하는 로직에서 fineDust-entity에서 점검 정보를 가져오는 로직으로 변경

+++ fine-dust/src/main/java/com/example/finedust/repository/CheckRepository.java
@@ -3,5 +3,7 @@ package com.example.finedust.repository;
 import com.example.finedust.entity.CheckEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CheckRepository extends JpaRepository<CheckEntity, Integer> {
+import java.util.List;
+
+public interface CheckRepository extends JpaRepository<CheckEntity, Long> {
 }
